Finished Chapter2
Some thoughts on the Chapter was finding out that the version used in the book is dated and there may be some translation issues later on. I will have to either work around them, downgrade my version of spaCy or find a different book to learn from. We'll see how the next chapter or 2 go.